ml-app/Jupyter-notebook/requirements.txt
jupyter
numpy
pandas
scikit-learn
matplotlib
seaborn
psycopg2
sqlalchemy
python-dotenv